User management (#11)
* first_commit

* crud

* fixe crud

* CRUD fix + Sign in (authentication + authorization)

* fix sign in

* implement email verification with token

* reset password

* test

* modification user

* modification user

* implement pdf

* scheduling + implementing sms

* refresh expired token (sheduling)

* implement pagination on users + commands

* fixed authorization and CRUD

* almost finished

* fix bugs

+++ src/main/java/tn/dalhia/services/implementations/CommandServiceImpl.java
@@ -1,11 +1,18 @@
 package tn.dalhia.services.implementations;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import javax.transaction.Transactional;
+
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
-
 
 import tn.dalhia.entities.Command;
 import tn.dalhia.entities.Product;
@@ -18,7 +25,7 @@ import tn.dalhia.request.CommandRequestModel;
 import tn.dalhia.response.ErrorMessages;
 import tn.dalhia.services.CommandService;
 import tn.dalhia.shared.dto.CommandDto;
-import tn.dalhia.shared.dto.Utils;
+import tn.dalhia.shared.tools.UtilsUser;
 
 @Service
 public class CommandServiceImpl implements CommandService{
@@ -33,14 +40,15 @@ public class CommandServiceImpl implements CommandService{
 	ProductRepository productRepo;
 	
 	@Autowired
-	Utils utils;
+    UtilsUser utils;
 	
 	@Override
 	@Transactional
-	public CommandDto createCommand(CommandRequestModel commandDetails, String id) {
+	public CommandDto createCommand(CommandRequestModel commandDetails, String id, Authentication authentification) {
 		User userEntity = userRepo.findByUserId(id);
 		if (userEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
 		if (commandDetails.getProducts() == null) throw new UserServiceException(ErrorMessages.MISSING_REQUIRED_FILED.getErrorMessage());
+		if(!utils.connectedUser(authentification,userEntity)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		
 		for (Product product : commandDetails.getProducts()) {
 			Product testProduct = productRepo.getById(product.getId());
@@ -67,7 +75,6 @@ public class CommandServiceImpl implements CommandService{
 //			commandEntity.setProducts(lP);
 //		}
 
-		
 		Command storedCommand = commandRepo.save(commandEntity);
 		ModelMapper modelMapper = new ModelMapper();
 		CommandDto returnValue = modelMapper.map(storedCommand, CommandDto.class);
@@ -75,9 +82,11 @@ public class CommandServiceImpl implements CommandService{
 	}
 
 	@Override
-	public CommandDto updateCommand(CommandRequestModel commandDetails, String id) {
+	public CommandDto updateCommand(CommandRequestModel commandDetails, String id, Authentication authentification) {
+		if(!utils.connectedUser(authentification,null)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		Command commandEntity = commandRepo.findByCommandId(id);
 		if (commandEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+
 		commandEntity.setCard(commandDetails.getCard());
 		commandEntity.setCode(commandDetails.getCode());
 		commandEntity.setEmail(commandDetails.getEmail());
@@ -91,18 +100,42 @@ public class CommandServiceImpl implements CommandService{
 	}
 
 	@Override
-	public CommandDto getCommandById(String id) {
+	public CommandDto getCommandById(String id, Authentication authentification) {
 		ModelMapper modelMapper = new ModelMapper();
 		Command command = commandRepo.findByCommandId(id);
+		if (command == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+		if(!utils.connectedUser(authentification,command.getUsers())) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		CommandDto returnValue = modelMapper.map(command,CommandDto.class);
 		return returnValue;
 	}
 
 	@Override
-	public void deleteCommand(String id) {
+	public void deleteCommand(String id, Authentication authentification) {
 		Command command = commandRepo.findByCommandId(id) ;
 		if (command == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+		if(!utils.connectedUser(authentification,command.getUsers())) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		commandRepo.delete(command);
+	}
+
+	@Override
+	public List<CommandDto> getCommandsPagination(int page, int limit, Authentication authentification) {
+		List<CommandDto> returnValue = new ArrayList<>();
+		if(!utils.connectedUser(authentification,null)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
+		if(page>0) page = page-1;
+		
+		Pageable pageableRequest = PageRequest.of(page, limit);
+		Page<Command>  commandsPage= commandRepo.findAll(pageableRequest);
+		
+		List<Command> commands = commandsPage.getContent();
+		
+		for (Command command : commands) {
+			
+			ModelMapper modelMapper = new ModelMapper();
+			CommandDto commandDto = modelMapper.map(command,CommandDto.class);
+			 //BeanUtils.copyProperties(command, commandDto);
+			 returnValue.add(commandDto);
+		}
+		return returnValue;
 	}
 
 }
