+++ src/main/java/tn/dalhia/services/implementations/UserServiceImpl.java
@@ -2,15 +2,19 @@ package tn.dalhia.services.implementations;
 
 import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
+import java.util.List;
 
 import javax.mail.MessagingException;
 import javax.mail.internet.MimeMessage;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.mail.javamail.MimeMessageHelper;
-
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -18,14 +22,15 @@ import org.springframework.stereotype.Service;
 
 import tn.dalhia.entities.PasswordResetTokenEntity;
 import tn.dalhia.entities.User;
+import tn.dalhia.entities.enumerations.Role;
 import tn.dalhia.exceptions.UserServiceException;
 import tn.dalhia.repositories.PasswordResetTokenRepository;
-
 import tn.dalhia.repositories.UserRepository;
 import tn.dalhia.response.ErrorMessages;
 import tn.dalhia.services.UserService;
 import tn.dalhia.shared.dto.UserDto;
-import tn.dalhia.shared.dto.Utils;
+import tn.dalhia.shared.tools.UtilsUser;
+
 
 
 
@@ -36,7 +41,7 @@ public class UserServiceImpl implements UserService {
 	UserRepository userRepo;
 	
 	@Autowired
-	Utils utils;
+    UtilsUser utils;
 	
 	@Autowired
 	BCryptPasswordEncoder bCryptPasswordEncoder;
@@ -61,6 +66,7 @@ public class UserServiceImpl implements UserService {
 		userEntity.setUserId(publicUserId);
 		userEntity.setEncryptedPaswword(bCryptPasswordEncoder.encode(user.getPassword()));
 		userEntity.setEmailVerificationToken(utils.generateEmailVerificationToken(publicUserId));
+		userEntity.setRole(Role.WOMAN);
 		User storedUserDetails =userRepo.save(userEntity);
 		
 		UserDto returnValue = new UserDto();
@@ -76,7 +82,7 @@ public class UserServiceImpl implements UserService {
 		String senderName = "Women App Team";
 		String mailContent = "<p>Dear " + user.getFirst_name() + user.getLast_name() + ",</p>";
 		mailContent += "<p> please check the link below to verify your email : </p>";
-		String verifyURL = "http://localhost:8089/pi/users/email-verification?token=" + user.getEmailVerificationToken();
+		String verifyURL = "http://localhost:8089/api/v1/users/email-verification?token=" + user.getEmailVerificationToken();
 		
 		mailContent += "<h2><a href=" + verifyURL + ">Verify your account</a></h2>";
 		
@@ -99,28 +105,36 @@ public class UserServiceImpl implements UserService {
 	public UserDto getUser(String email) {
 		User userEntity = userRepo.findByEmail(email);
 		
-		if (userEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
-		UserDto returnValue = new UserDto();
-		BeanUtils.copyProperties(userEntity,returnValue);
-		return returnValue;
-
+			if (userEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+			UserDto returnValue = new UserDto();
+			BeanUtils.copyProperties(userEntity,returnValue);
+			return returnValue;
+		
+		
+		
 	}
 
 	@Override
-	public UserDto getUserByUserId(String userId) {
+	public UserDto getUserByUserId(String userId, Authentication authentification) {
 		UserDto returnValue = new UserDto();
 		User userEntity = userRepo.findByUserId(userId);
+		if(!utils.connectedUser(authentification,userEntity)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
+
 		if (userEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
 		
 		BeanUtils.copyProperties(userEntity,returnValue);
 		return returnValue;
+
+	
+		 
 	}
 
 	@Override
-	public UserDto updateUser(String id, UserDto userDto) {
+	public UserDto updateUser(String id, UserDto userDto, Authentication authentification) {
 		UserDto returnValue = new UserDto();
 		User userEntity = userRepo.findByUserId(id);
 		
+		if(!utils.connectedUser(authentification,userEntity)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		if (userEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage()); //Exception eli aamlneha ahna
 		
 			userEntity.setAddress(userDto.getAddress());
@@ -138,9 +152,9 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public void deleteUser(String userId) {
+	public void deleteUser(String userId,Authentication authentification) {
 		User userEntity = userRepo.findByUserId(userId);
-		
+		if(!utils.connectedUser(authentification,userEntity)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
 		if (userEntity == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
 		
 		userRepo.delete(userEntity);
@@ -166,7 +180,7 @@ public class UserServiceImpl implements UserService {
 		User userEntity = userRepo.findUserByEmailVerificationToken(token);
 		
 		if(userEntity!=null) {
-			boolean hastokenExpired = Utils.hastokenExpired(token);
+			boolean hastokenExpired = UtilsUser.hastokenExpired(token);
 			if (!hastokenExpired) { // netfakdou token expired wla not expired yet
 				userEntity.setEmailVerificationToken(null);
 				userEntity.setEmailVerificationStatus(Boolean.TRUE);
@@ -186,7 +200,7 @@ public class UserServiceImpl implements UserService {
 			return returnValue;
 		}
 		
-		String token =  new Utils().generatePasswordResetToken(userEntity.getUserId()); //wala nrodhha static f utils
+		String token =  new UtilsUser().generatePasswordResetToken(userEntity.getUserId()); //wala nrodhha static f utils
 		
 		PasswordResetTokenEntity passwordResetTokenEntity = new PasswordResetTokenEntity();
 		passwordResetTokenEntity.setToken(token);
@@ -231,11 +245,13 @@ public class UserServiceImpl implements UserService {
 	public boolean resetPassword(String token, String password) {
 		boolean returnValue = false;
 		
-		if(Utils.hastokenExpired(token)){
+		PasswordResetTokenEntity passwordResetTokenEntity =passwordResetTokenRepository.findByToken(token);
+		if(UtilsUser.hastokenExpired2(token)){
+			passwordResetTokenRepository.delete(passwordResetTokenEntity);
 			return returnValue;
 		}
 		
-		PasswordResetTokenEntity passwordResetTokenEntity =passwordResetTokenRepository.findByToken(token);
+
 		
 		if(passwordResetTokenEntity==null){
 			return returnValue;
@@ -261,4 +277,52 @@ public class UserServiceImpl implements UserService {
 	
 	}
 
+	@Override
+	public List<User> getUsers() {
+
+		List<User> userEntities = userRepo.findAll();
+		if (userEntities == null) throw new UserServiceException(ErrorMessages.NO_RECORD_FOUND.getErrorMessage());
+		
+		
+		return userEntities;
+	}
+
+	@Override
+	//@Scheduled(fixedRate = 60000/60)
+	public void checktokenExpiration() throws UnsupportedEncodingException, MessagingException {
+		List<User> userEntities = userRepo.findAll();
+		
+		for(User user : userEntities ) {
+		if(UtilsUser.hastokenExpired2(user.getEmailVerificationToken())){
+			user.setEmailVerificationToken(utils.generateEmailVerificationToken(user.getUserId()));
+			User updatedUserDetails = userRepo.save(user);
+			
+			UserDto returnValue = new UserDto();
+			BeanUtils.copyProperties(updatedUserDetails,returnValue);
+			
+			sendVerificationEmail(returnValue);
+		}
+		}
+	}
+
+	@Override
+	public List<UserDto> getUsersPagination(int page, int limit, Authentication authentification) {
+		
+		if(!utils.connectedUser(authentification,null)) throw new UserServiceException(ErrorMessages.SECURITY_ERROR.getErrorMessage());
+		List<UserDto> returnValue = new ArrayList<>();
+		
+		if(page>0) page = page-1;
+		
+		Pageable pageableRequest = PageRequest.of(page, limit);
+		Page<User>  usersPage= userRepo.findAll(pageableRequest);
+		
+		List<User> users = usersPage.getContent();
+		
+		for (User user : users) {
+			 UserDto userDto = new UserDto();
+			 BeanUtils.copyProperties(user, userDto);
+			 returnValue.add(userDto);
+		}
+		return returnValue;
+	}
 }
